requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
langchain==0.3.26
flask==3.1.1 
sentence-transformers==4.1.0
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-pinecone==0.2.8 
langchain-openai==0.3.24
langchain-community==0.3.26
langchain-google-genai==0.0.8
google-genai

Werkzeug
email-validator
-e .